test: enable CodeInspection-integ-test
