+++ test/SpecRunner.html
@@ -63,7 +63,7 @@
 
 </head>
 
-<body>
+<body style="user-select: text !important;">
     <div class="navbar navbar-fixed-top">
       <div class="navbar-inner">
         <div class="container">
