+++ test/UnitTestSuite.js
@@ -76,8 +76,8 @@ define(function (require, exports, module) {
     require("spec/ViewFactory-test");
     require("spec/ViewUtils-test");
     require("spec/XMLUtils-test");
+    require("spec/CodeInspection-integ-test");
     // require("spec/CodeHint-test");
-    // require("spec/CodeInspection-test");
     // require("spec/CSSInlineEdit-test");
     // require("spec/DocumentCommandHandlers-test");
     // require("spec/DocumentManager-test");
